+++ app/lib/folio/application_cell.rb
@@ -97,11 +97,11 @@ class Folio::ApplicationCell < Cell::ViewModel
     end
   end
 
-  def render_component(component)
+  def render_view_component(component)
     if context[:view]
       context[:view].render(component)
     else
-      fail "Missing context[:view] - cannot render_component"
+      fail "Missing context[:view] - cannot render_view_component"
     end
   end
 end